refactor: centralize user command logic
